chore: decouple rpc client code from server code
