+++ crates/ursa/src/ursa/rpc_commands.rs
@@ -1,7 +1,9 @@
 use structopt::StructOpt;
 use tracing::{error, info};
-use ursa_rpc_client::functions::{get_file, put_file};
-use ursa_rpc_server::api::{NetworkGetFileParams, NetworkPutFileParams};
+use ursa_rpc_client::{
+    functions::{get_file, put_file},
+    NetworkGetFileParams, NetworkPutFileParams,
+};
 
 #[derive(Debug, StructOpt)]
 pub enum RpcCommands {
